+++ src/runtime/oflux_exercise.cpp
@@ -49,7 +49,7 @@ main(int argc, char * argv[])
 	}
 	char * max_nsec_wait_str = getenv("EXERCISE_SLEEP");
 	if(max_nsec_wait_str) {
-		oflux::flow::exercise::max_nsec_wait = atoi(max_nsec_wait_str);
+		oflux::flow::exercise::max_nsec_wait = atol(max_nsec_wait_str);
 	}
 	int init_threads = 0;
 	char * init_threads_str = getenv("EXERCISE_THREADS");
@@ -71,7 +71,7 @@ main(int argc, char * argv[])
 		, init_atomic_maps
 		};
 	oflux::logging::toStream(std::cout); 
-	oflux_log_info(" exercise::max_nsec_wait is %d\n", oflux::flow::exercise::max_nsec_wait);
+	oflux_log_info(" exercise::max_nsec_wait is %ld\n", oflux::flow::exercise::max_nsec_wait);
 	oflux::EnvironmentVar env(oflux::runtime::Factory::classic);
 	oflux::flow::ExerciseFunctionMaps ffmaps(env.runtime_number == 4);
 	atomic_set = ffmaps.atomic_set();
